Change BoundingBox and CardSizesContainer attributes from float to int

+++ KC.Detector/Models/BoundingBox.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 @dataclass
 class BoundingBox:
-    x : float
-    y : float
-    w : float
-    h : float
+    x : int
+    y : int
+    w : int
+    h : int
